+++ client/src/components/home/Main.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Main = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Main
